StatusEffectManager added, along with StatusEffects.txt for loading them in.  Currently have all effects except Stats up/down; those are going to require a little more work.
The Manager is a singleton, and has a GetStatus(std::string) method that returns the address of the status effect.  They are designed to have copies made and attached to affected characters, as each instance of a status effect requires certain information unique to the circumstance.  This is especially true with the Stats up/down effects.  

TurnManager will be the middle man on this, getting all the math assigned appropriately before adding the status to the character.  Will be adding a function in Turnmanager to specifically handle Abilities being used.  Things are getting exciting.